+++ wxc_sdk/as_api.py
@@ -55,16 +55,16 @@ __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'A
            'AsOutgoingPermissionsApi', 'AsPSTNApi', 'AsPagingApi', 'AsPeopleApi', 'AsPersonForwardingApi',
            'AsPersonSettingsApi', 'AsPersonSettingsApiChild', 'AsPersonalAssistantApi', 'AsPlayListApi',
            'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPriorityAlertApi', 'AsPrivacyApi',
-           'AsPrivateNetworkConnectApi', 'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi',
-           'AsReceptionistContactsDirectoryApi', 'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRolesApi',
-           'AsRoomTabsApi', 'AsRoomsApi', 'AsRouteGroupApi', 'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2GroupsApi',
-           'AsSCIM2UsersApi', 'AsScheduleApi', 'AsScimApiChild', 'AsScimV2Api', 'AsSelectiveAcceptApi',
-           'AsSelectiveForwardApi', 'AsSelectiveRejectApi', 'AsSequentialRingApi', 'AsSimRingApi',
-           'AsSingleNumberReachApi', 'AsStatusAPI', 'AsSupervisorApi', 'AsTeamMembershipsApi', 'AsTeamsApi',
-           'AsTelephonyApi', 'AsTelephonyDevicesApi', 'AsTelephonyLocationApi', 'AsTransferNumbersApi',
-           'AsTranslationPatternsApi', 'AsTrunkApi', 'AsUpdateRoutingPrefixJobsApi', 'AsVirtualExtensionsApi',
-           'AsVirtualLinesApi', 'AsVoiceMessagingApi', 'AsVoicePortalApi', 'AsVoicemailApi', 'AsVoicemailGroupsApi',
-           'AsVoicemailRulesApi', 'AsWebexSimpleApi', 'AsWebhookApi', 'AsWorkspaceDevicesApi',
+           'AsPrivateNetworkConnectApi', 'AsPushToTalkApi', 'AsQueueCallRecordingSettingsApi',
+           'AsRebuildPhonesJobsApi', 'AsReceptionistApi', 'AsReceptionistContactsDirectoryApi', 'AsRecordingsApi',
+           'AsReportsApi', 'AsRestSession', 'AsRolesApi', 'AsRoomTabsApi', 'AsRoomsApi', 'AsRouteGroupApi',
+           'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2GroupsApi', 'AsSCIM2UsersApi', 'AsScheduleApi',
+           'AsScimApiChild', 'AsScimV2Api', 'AsSelectiveAcceptApi', 'AsSelectiveForwardApi', 'AsSelectiveRejectApi',
+           'AsSequentialRingApi', 'AsSimRingApi', 'AsSingleNumberReachApi', 'AsStatusAPI', 'AsSupervisorApi',
+           'AsTeamMembershipsApi', 'AsTeamsApi', 'AsTelephonyApi', 'AsTelephonyDevicesApi', 'AsTelephonyLocationApi',
+           'AsTransferNumbersApi', 'AsTranslationPatternsApi', 'AsTrunkApi', 'AsUpdateRoutingPrefixJobsApi',
+           'AsVirtualExtensionsApi', 'AsVirtualLinesApi', 'AsVoiceMessagingApi', 'AsVoicePortalApi', 'AsVoicemailApi',
+           'AsVoicemailGroupsApi', 'AsVoicemailRulesApi', 'AsWebexSimpleApi', 'AsWebhookApi', 'AsWorkspaceDevicesApi',
            'AsWorkspaceLocationApi', 'AsWorkspaceLocationFloorApi', 'AsWorkspaceNumbersApi',
            'AsWorkspacePersonalizationApi', 'AsWorkspaceSettingsApi', 'AsWorkspacesApi', 'AsWrapupReasonApi',
            'AsXApi']
@@ -18129,6 +18129,98 @@ class AsConferenceControlsApi(AsApiChild, base='telephony/conference'):
         await super().post(url, json=body)
 
 
+class AsQueueCallRecordingSettingsApi(AsApiChild, base='telephony/config/locations'):
+    """
+    Queue Call Recording Settings
+
+    Queue Call Settings supports modifying Webex Calling settings for a specific queue.
+
+    Viewing Queue recording settings requires a full, user, or read-only administrator or location administrator auth
+    token with a scope of `spark-admin:people_read` or, for select APIs, a user auth token with `spark:people_read`
+    scope can be used by a person to read their own settings.
+
+    Configuring Queue recording settings requires a full or user administrator or location administrator auth token
+    with the `spark-admin:people_write` scope or, for select APIs, a user auth token with `spark:people_write` scope
+    can be used by a person to update their own settings.
+
+    Call Queue Recording Settings API access can be restricted via Control Hub by a full administrator. Restricting
+    access causes the APIs to throw a `403 Access Forbidden` error.
+
+    See details about `features available by license type for Webex Calling
+    <https://help.webex.com/en-us/article/n1qbbp7/Features-available-by-license-type-for-Webex-Calling>`_.
+    """
+
+    async def read(self, location_id: str, queue_id: str,
+             org_id: str = None) -> CallRecordingSetting:
+        """
+        Read Queue Call Recording Settings for a Queue
+
+        Retrieve a queue's Call Recording settings.
+
+        The Call Recording feature provides a hosted mechanism to record the calls placed and received on the Carrier
+        platform for replay and archival. This feature is helpful for quality assurance, security, training, and more.
+
+        This API requires a full or user administrator or location administrator auth token with the
+        `spark-admin:people_read` scope.
+
+        A person with a Webex Calling Standard license is eligible for the Call Recording
+        feature only when the Call Recording vendor is Webex.
+
+        :param location_id: Unique identifier for the location.
+        :type location_id: str
+        :param queue_id: Unique identifier for the queue.
+        :type queue_id: str
+        :param org_id: ID of the organization in which the queue resides. Only admin users of another organization
+            (such as partners) may use this parameter as the default is the same organization as the token used to
+            access API.
+        :type org_id: str
+        :rtype: :class:`CallRecordingSetting`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep(f'{location_id}/queues/{queue_id}/cxEssentials/callRecordings')
+        data = await super().get(url, params=params)
+        r = CallRecordingSetting.model_validate(data)
+        return r
+
+    async def configure(self, location_id: str, queue_id: str,
+                  recording: CallRecordingSetting,
+                  org_id: str = None):
+        """
+        Configure Queue Call Recording Settings for a Queue
+
+        Configure a queue's Call Recording settings.
+
+        The Call Recording feature provides a hosted mechanism to record the calls placed and received on the Carrier
+        platform for replay and archival. This feature is helpful for quality assurance, security, training, and more.
+
+        This API requires a full or user administrator or location administrator auth token with the
+        `spark-admin:people_write` scope.
+
+        <div><Callout type="warning">A person with a Webex Calling Standard license is eligible for the Call Recording
+        feature only when the Call Recording vendor is Webex.</Callout></div>
+
+        :param location_id: Unique identifier for the location.
+        :type location_id: str
+        :param queue_id: Unique identifier for the queue.
+        :type queue_id: str
+        :param recording: the new recording settings
+        :type recording: CallRecordingSetting
+        :param org_id: ID of the organization in which the queue resides. Only admin users of another organization
+            (such as partners) may use this parameter as the default is the same organization as the token used to
+            access API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = recording.update()
+        url = self.ep(f'{location_id}/queues/{queue_id}/cxEssentials/callRecordings')
+        await super().put(url, params=params, json=body)
+
+
 class AsWrapupReasonApi(AsApiChild, base='telephony/config'):
     """
     Wrap up reasons API
@@ -18451,10 +18543,12 @@ class AsCustomerExperienceEssentialsApi(AsApiChild, base='telephony/config'):
     Modifying the customer Experience Essentials APIs requires a full or device administrator auth token with a scope
     of `spark-admin:telephony_config_write`.
     """
+    callqueue_recording: AsQueueCallRecordingSettingsApi
     wrapup_reasons: AsWrapupReasonApi
 
     def __init__(self, session: AsRestSession):
         super().__init__(session=session)
+        self.callqueue_recording = AsQueueCallRecordingSettingsApi(session=session)
         self.wrapup_reasons = AsWrapupReasonApi(session=session)
 
     async def get_screen_pop_configuration(self, location_id: str = None,
@@ -26174,7 +26268,7 @@ class AsLocationNumbersApi(AsApiChild, base='telephony/config/locations'):
     A partner administrator can retrieve or change settings in a customer's organization using the optional `orgId`
     query parameter.
     """
-    
+
     def _url(self, location_id: str, path: str = None):
         """